[TE] move threadlocal to auth filter
